New model flavor, and adding batchnorm
Baseline has been established pretty well, and now we're moving
ahead to explore some different aspects.

BatchNorm is now included in the DecoderBlock, so that we're running
batch norm between layers.

Also started training a TransposeConvolution model that replaces
the base decoder: instead of doing an upsample + convolution, we're
using a transpose convolution layer to perform the upsampling.