Beefed up the architecture in general and bugfix
The encoder now has a 7x7 convolution as its first operation in the
hopes of capturing more long range parts of the image for the
reconstruction. The decoder, I realized, was also incredibly lightweight
compared to the decoder (most of the capacity was in the first fully
connected layer). I've now made the kernel sizes for each decoder layer
larger, in order to bump up its reconstruction ability.

It also turned out that the final layer of the decoder was still adding
an activation function to the output, which was unwanted since we want
logits out. This has now been fixed.