Denoising deterministic autoencoder training well
There were two big things that was stopping learning: first,
BCEwithLogits was producing overly blurry images for whatever
reason - initially I thought this was because the noisy images contained
negative numbers (which was the second reason) but it turned out to not
be the case.

Second was the fact that the Gaussian noised images were
normalizing to a maximum of one, but the negative values remained which
are suboptimal. The pipeline is now modified to be AddNoise -> Normalize
-> ToTensor, where the Normalize now resembles `MinMaxScaler` from
`sklearn`, i.e. pixel values are clamped to [0,1].