Major refactoring of argparsing
All of the model specific argument parsing has been shifted out of the
training scripts. There is a `common_hyperparameters` function in `base`
that defines, as the name suggests, hyperparameters that are common to
all models (e.g. learning rate, number of channels, etc), and a static
method has been added to each model to include model specific arguments
per the recommended PyTorch lightning style. Nominally, for VAE this is
just the beta coefficient but could be extended to other things.

The `train_master.py` script should now encompass all training, and
we'll gradually move over to that script. This script includes the
argparse for the Trainer specific arguments, and combines with the model
parsing.