Added PixelVAE and DropBlock
For PixelVAE, this literally VAE with a PixelCNN tacked onto the output
of the original decoder. The idea behind this is that first CNN layers
produce the upscaled image, and the PixelCNN does a final improvement of
the image.

I've also replaced `nn.Dropout` in the models with `DropBlock2d`, which
is supposed to be a better adapted method for dropout in CNNs.