Changed the noise generation in the pipeline
Instead of one or the other, we now combine Gaussian and Poisson noise
to the image. The first is simulated with a uniform variance sampled
between a range, and the second uses `skimage`'s routine - the result is
combined to give the joint image.

Will need to characterize this properly, i.e. measure the SNR coming
out.