Added a new method of dataset generation
Instead of doing it stochastically, just generate a 3D grid of
parameters for every possible combination of beta, mu, and sigma over a
range.

Unfortunately the way it is designed with `np.mgrid` we can't actually
easily specify _how_ many images we want ultimately, and this is hard
coded. Regardless, the idea would be to simply have your dataset become
less and less sparse if we want to train a continuous function.