Added SVAEGAN
This model is a minimally modified VAEGAN, where instead of the usual
architecture for the discriminator, we reuse the encoder from the
autoencoder. The idea is that the encoder learns to extract features
from an image simultaenously to reconstruct it, and it identify if it
was a fake image or not. I don't know if this is a novel architecture,
but at the least I hope this beefs up the encoder such that the decoder
doesn't start to overfit.

I also don't know if the training dynamics are stable for this...Time
will tell...